`refactor(home, auth): Clean up and optimize screen components and data sources`
- Removed unused platform configurations from .metadata.
- Simplified LoginScreen by removing the optional verificationCode parameter.
- Updated HomeRemoteDataSource to include image_url and limit the number of deals fetched.
- Replaced FeaturedPros widget with ServicesCarousel and PromotionsBanner in HomeScreen.
- Deleted obsolete featured_pros.dart and special_offers.dart files to streamline widget structure.
- Adjusted app routing for cleaner navigation and removed unnecessary imports.
